move address tests and subnetwork tests
